Export memmap from DNA port E2 test, use memmap in binary
