--- firmware-binaries/sim-tests/dna_port_e2_test/memory.x
@@ -1,18 +1,0 @@
-/*
-SPDX-FileCopyrightText: 2024 Google LLC
-
-SPDX-License-Identifier: CC0-1.0
-*/
-
-MEMORY
-{
-  IMEM : ORIGIN = 0x80000000, LENGTH = 4K
-  DMEM : ORIGIN = 0x20000000, LENGTH = 4K
-}
-
-REGION_ALIAS("REGION_TEXT", IMEM);
-REGION_ALIAS("REGION_RODATA", DMEM);
-REGION_ALIAS("REGION_DATA", DMEM);
-REGION_ALIAS("REGION_BSS", DMEM);
-REGION_ALIAS("REGION_HEAP", DMEM);
-REGION_ALIAS("REGION_STACK", DMEM);
