+++ firmware-binaries/sim-tests/dna_port_e2_test/src/main.rs
@@ -6,15 +6,17 @@
 
 use ufmt::uwriteln;
 
-use bittide_hal::shared_devices::{Dna, Uart};
+use bittide_hal::hals::dna_port_e2_test::DeviceInstances;
 #[cfg(not(test))]
 use riscv_rt::entry;
+
+const DEVICES: DeviceInstances = unsafe { DeviceInstances::new() };
 
 #[cfg_attr(not(test), entry)]
 fn main() -> ! {
     // Initialize peripherals.
-    let mut uart = unsafe { Uart::new(0x4000_0000 as *mut u8) };
-    let dna = unsafe { Dna::new(0x6000_0000 as *mut u8) };
+    let mut uart = DEVICES.uart;
+    let dna = DEVICES.dna;
     let mut dna_arr = [0; 16];
     dna_arr[0..12].copy_from_slice(&dna.dna());
     let dna_val = u128::from_le_bytes(dna_arr);
